[CCT] Deprecate some of the CCT APIs
Based on an analysis that was performed in Q2/Q3 2023 it was decided
that some of the CCT APIs will be deprecated. See go/cct-specific-apis-deprecation-2023
for more information.

As the first step the APIs will be deprecated under a feature flag that
will act as a kill switch in case we discover unexpected usages and
after the APIs have been disabled for a while in Stable a second step
will follow that will remove the code gated under the feature flag.

Low-Coverage-Reason: only deprecating part of the functionality that is currently not covered by tests.
Bug: 1474699
Change-Id: Ia635b82dbab184b94ab75655bf1006b392a3485b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4798867
Reviewed-by: Jinsuk Kim <jinsukkim@chromium.org>
Commit-Queue: Kevin Grosu <kgrosu@google.com>
Cr-Commit-Position: refs/heads/main@{#1189038}