Suppress toasts while the Welcome Tour is in progress.
Suppress toasts while the Welcome Tour is in progress. Toasts that are
suppressed will not be restored when the Welcome Tour ends as toasts are
tied to a specific context that has likely changed.

Note that this CL also ensures that the Welcome Tour is aborted when Shell is destroying so that the tour does not outlive its dependencies.

Bug: b:277091624
Change-Id: Ibaa318a55546aefac929e1456f615a04bc278cc6
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4807375
Reviewed-by: Angus McLean <angusmclean@google.com>
Commit-Queue: David Black <dmblack@google.com>
Cr-Commit-Position: refs/heads/main@{#1189065}