Use powershell rather than cmd for CRD custom actions
I tested a nightly build with my previous change and noticed that
cmd windows were still popping up during installation. It looks like
this is because I was using 'cmd /c' in the 'quiet' custom action.

I tested the previous change and did not see CMD windows but I must
have been testing in such a way that the actions were not run as I
can reproduce that problem with a clean package install now.

The fix is to use powershell instead of cmd for the actions since it
allows passing flags which cause it to run w/o a window. The only
quirky change is that the 'Remove-item' command wants to recursively
delete the directory too which we may not want so I'm using a
dance to get the symlink and 'delete' it rather than use Remove-Item.

Bug: b/297319930
Change-Id: Ie9ea727bc36455fc972498d58b5771bd68b6a228
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4818071
Auto-Submit: Joe Downing <joedow@chromium.org>
Reviewed-by: Jamie Walch <jamiewalch@chromium.org>
Commit-Queue: Joe Downing <joedow@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1189069}