FSA: Implement FileSystemObserver.unobserve()
Good-enough implementation of the unobserve() method. It's possible for
events to be relayed to the renderer in the time between when the
Unobserve() IPC is posted and when the observation is removed in the
browser process.  A more "correct" implementation will require more
invasive changes such as queueing in the renderer - to be left for a
follow-up CL.

Bug: 1019297
Change-Id: I4cba8fbe8a5fe79be885c8ab6d5d348ddd1b1866
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4814709
Commit-Queue: Austin Sullivan <asully@chromium.org>
Reviewed-by: Evan Stade <estade@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1189077}