Notify GC about external memory allocated on font objects
GC is not aware of the amount of memory allocated for the font objects.
Therefore, when a lot of font objects are created, for example, during
heavy animation of variable web fonts, see [0], memory is growing
unboundedly, without GC kicking in, until the tab crashes.

We have two types of large allocations related to the font objects:
First, when we decode font resource, we allocate decoded resource size
of bytes for the new instance of SkData [1].

Second, during the animation of variable web fonts a lot of
SimpleFontData objects are created, one per each animation frame. Each
time we create new SimpleFontData object, we will invoke
SkFont::getWidthsBounds() to get the width for zero and space glyph,
compare [2] and [3]. Then, in ScalerContext_FreeType::generateMetrics(),
Skia will ask FreeType to get glyphs metrics by invoking
af_face_globals_get_metrics, see [4]. In af_face_globals_get_metrics,
will allocate style_metrics_size bytes of memory on the metrics, compare
[5].

To prevent the unbounded growth of memory, we need to inform GC about
external memory allocated on font objects. This CL notifies GC about the
amount of memory allocated on Font objects when the font resource is
decoded and when SimpleFontData objects are created. In the first case
we are using the size of the network resource as the amount of memory
allocated on the font objects. In the second case we use max value of
style_metrics_size [6] for that purpose.

Test: manually tested on
third_party/blink/web_tests/http/tests/ManualTests/fonts/leakyvar.html
before the change memory would grow up to 15GB and then the tab would
crash, after the change memory stays under 350MB.

[0] https://bugs.chromium.org/p/chromium/issues/detail?id=778352#c59
[1] https://source.chromium.org/chromium/chromium/src/+/main:third_party/blink/renderer/core/css/remote_font_face_source.cc?q=RemoteFontFaceSource::NotifyFinished&ss=chromium%2Fchromium%2Fsrc
[2] https://source.chromium.org/chromium/chromium/src/+/main:third_party/blink/renderer/platform/fonts/simple_font_data.cc?q=WidthForGlyph%5C(space_glyph_%5C)&ss=chromium%2Fchromium%2Fsrc
[3] https://source.chromium.org/chromium/chromium/src/+/main:third_party/blink/renderer/platform/fonts/simple_font_data.cc?q=WidthForGlyph%5C(zero_glyph_%5C)&ss=chromium%2Fchromium%2Fsrc
[4] https://source.chromium.org/chromium/chromium/src/+/main:third_party/freetype/src/src/autofit/afglobal.c?q=af_face_globals_get_metrics&ss=chromium%2Fchromium%2Fsrc
[5] https://source.chromium.org/chromium/chromium/src/+/main:third_party/freetype/src/src/autofit/afglobal.c?q=style_metrics_size&ss=chromium%2Fchromium%2Fsrc
[6] https://source.chromium.org/chromium/chromium/src/+/main:third_party/freetype/src/src/autofit/aftypes.h?q=style_metrics_size&ss=chromium%2Fchromium%2Fsrc

Bug: 1429581
Change-Id: I0fb90931c8a011d495dea113a586c0e51ec16309
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4717485
Reviewed-by: Dominik Röttsches <drott@chromium.org>
Commit-Queue: Munira Tursunova <moonira@google.com>
Cr-Commit-Position: refs/heads/main@{#1189092}