WebUI Omnibox: Open autocomplete match from cr-realbox-match
This CL moves the call to open an autocomplete match from the NTP into
cr-realbox-match so that it can be called for all shared usages of
the cr-realbox-dropdown. This ultimately enables clicking on the
WebUI omnibox to navigate to a page.

Bug: 1451051
Change-Id: I3cfc9b56108b3f137054c6f4d4e8686dbc6a48f8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4807426
Commit-Queue: John Lee <johntlee@chromium.org>
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Reviewed-by: Moe Ahmadi <mahmadi@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1189096}