[GM3] Set ThemeID for UserColor themes
The appearance section of the settings page is currently dependent on
ThemeID to know whether the theme is default or not. This experience
probably overall needs an update, but settings ThemeID for GM3 color and
checking for it at least sets the experience back to matching GM2.

Without this change, the "Reset to default"/"Reset to classic" button
does not appear in chrome://settings/appearance with a GM3 theme.

I did not add to the browser test since the browser test for the button
showing still works, since the button existing depends more on ThemeID
being set at all. The value is only for the "Chrome Colors" text.

Bug: 1475611
Change-Id: I574db1e3740355c1bfd381c51fc214387fb3bcaa
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4813446
Reviewed-by: Peter Kasting <pkasting@chromium.org>
Commit-Queue: Riley Tatum <rtatum@google.com>
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Reviewed-by: John Lee <johntlee@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1189097}