[SecureDLs] Update more tests to use secure downloads
This CL updates some download tests to use HTTPS-based downloads instead
of HTTP. This is in preparation for warnings on insecure warnings (which
causes these tests to fail).

As part of this change, this CL updates tests to use non-public
hostnames a.test and b.test instead of registrable .com names. This
allows the HTTPS server to offer a valid certificate to Chrome.

Bug: 1352598
Change-Id: I0cafea38102d663ee7ed63b5340fb91c879f75ad
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4808888
Reviewed-by: Min Qin <qinmin@chromium.org>
Auto-Submit: Joe DeBlasio <jdeblasio@chromium.org>
Commit-Queue: Min Qin <qinmin@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1189138}