[iOS] Improve top-of-feed offset function
The changes in https://crrev.com/c/4730054 improved the dimensions
functions for the NTP, making it easier to calculate offsets that
depend on the heights of content above the feed.

This CL fixes crbug/1476025 and greatly simplifies the function that
returns the offset when scrolled to the top of the feed.

Fixed: 1476025
Change-Id: Id303591f5e1fb905288fcaf82d068cfa55f77dfa
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4814366
Reviewed-by: Chris Lu <thegreenfrog@chromium.org>
Reviewed-by: Scott Yoder <scottyoder@google.com>
Commit-Queue: Adam Arcaro <adamta@google.com>
Cr-Commit-Position: refs/heads/main@{#1189149}