[Default Nav Transition] Let surface pick up the exact CopyOutRequest.
Before this change the arrived exact request will be queued up the
CompositorFrameSinkSupport before it is picked up by the targeted
surface. This CL makes sure that the exact request is picked up upon
arrival.

In a cross-renderer navigation, the old document will be unloaded in the
old renderer, thus tearing down the old CompositorFrameSink. If we then
submit a copy request, that request will have no
CompositorFrameSinkSupport to go, even though the targeted surface is
still alive.

Bug: 1414172, 1465547
Low-Coverage-Reason: the feature CL on top of this will have full tests.
Change-Id: I56578c7bf82277661ed35c81b168606663827f6e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4790841
Reviewed-by: Khushal Sagar <khushalsagar@chromium.org>
Reviewed-by: Jonathan Ross <jonross@chromium.org>
Commit-Queue: William Liu <liuwilliam@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1189157}