Speculative fix for uptick in AccessibilityFatalError calls
Partial revert for CL:4753355. This is in the right time frame, and
seems plausible. It's possible that the destruction of the
AXTreeManager comes while there is a serialization in flight, and
that this code somehow makes that worse.

Bug: 1471373
Change-Id: Ifebc5b68067c1fa7ddaf526830f6afc576bedf99
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4811427
Reviewed-by: Chris Harrelson <chrishtr@chromium.org>
Commit-Queue: Aaron Leventhal <aleventhal@chromium.org>
Auto-Submit: Aaron Leventhal <aleventhal@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1189175}