accelerators: Rework launcher handling decision into a state machine
Previously on ChromeOS, the decision whether to toggle the launcher
state was handled by the AcceleratorHistory component which kept track
of all currently pressed keys.

The issue with this component is that it was at a level in the stack where it could potentially miss some real key presses and releases.

This could result in the launcher not being able to be toggled until a
user presses or releases the exact key which was not recorded by
AcceleratorHistory which lead to a lot of frustration as the key was
difficult to track down.

To detect when "No keys are pressed", a new API was added to the
InputController which notes when any keys are physically pressed on any
keyboard connected to the system.

Bug: b/285970320
Change-Id: Id64bb198a9385c68690cb33d4c5a71b9bd28e866
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4730472
Reviewed-by: Scott Violet <sky@chromium.org>
Reviewed-by: Sean O'Brien <seobrien@chromium.org>
Reviewed-by: Zentaro Kavanagh <zentaro@chromium.org>
Commit-Queue: David Padlipsky <dpad@google.com>
Cr-Commit-Position: refs/heads/main@{#1189193}