Add clientAutomatedTest field to request
Adding a new field to the client requests controlled with a flag, this
field will signal the autopush reporting server that the events are
being sent from a client automated test (tast). This will help us
disable some checks and features on the server side that might be
making the tests fail (eg: the device deprovisioned check).

BUG=b:297888130
TEST= deployed and tested locally with lab device.

Change-Id: I0c2ea909c575d3cdf29685df532b5967d0f29bcd
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4817971
Reviewed-by: Hong Xu <xuhong@google.com>
Commit-Queue: Alberto Juarez <albertojuarez@google.com>
Cr-Commit-Position: refs/heads/main@{#1189201}