Log Reading mode settings changes from the webui toolbar
These are existing metrics that are currently logged in
ReadAnythingController[1]. Since the webui toolbar is behind a flag,
there will be negligible public usage of this until release, and at
that point we will remove the logging from ReadAnythingController.

This is part of the broader effort to refactor the toolbar from
views to webui go/read-mode-toolbar-redesign.

[1] https://source.chromium.org/chromium/chromium/src/+/main:chrome/browser/ui/views/side_panel/read_anything/read_anything_controller.cc;bpv=0;bpt=0

Bug: 1465029
Change-Id: I334719877ed0fbc6e659cb97b987a47c4cb99ea2
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4810529
Reviewed-by: Tim <tjudkins@chromium.org>
Commit-Queue: Kristi Saney <kristislee@google.com>
Reviewed-by: Abigail Klein <abigailbklein@google.com>
Cr-Commit-Position: refs/heads/main@{#1189228}