[wallets] Make clientId and configURL optional for Wallets
In this CL, we make clientId and configURL optional fields for
wallets, while keeping them required by federated identity providers.

We accomplish that by removing the WebIDL "required" property and
check at runtime if they are available for federated providers, while
skipping that check for wallet providers.

With that, we make this change backwards compatible while allowing
non-federated providers to avoid passing unnecessary fields.

Note: this CL changes a bit how we are counting  UseCounter::Count(context, WebFeature::kFedCm) in that it can over-count invocations that would previously be Type errors before counting to post-counting. Given that the Wallet API is behind a flag, we don't expect this to have any effect in practice.

Change-Id: I8b66fac34fe6fd9f1c3158b0787d8ea81e2f2212
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4808508
Reviewed-by: Nicolás Peña <npm@chromium.org>
Commit-Queue: Sam Goto <goto@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1189243}