[CVV Storage] Implement ApplyDisableSyncChanges in the sync bridge
This API will be triggered when the sync is disabled and we want
to clear out all the Server CVC data from local database.

DD- go/pay-autofill-cvv-storage-chrome

error are tested here. Testing the error line is not useful and
would be adding unnecessary tests as stated in the
doc(go/bypass-low-coverage-check)

Low-Coverage-Reason: All the lines except for the line to report
Bug: 1458044
Change-Id: I1e2aa28836aebe21beb97b05ee89f6da7508fad4
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4799838
Reviewed-by: Siyu An <siyua@chromium.org>
Reviewed-by: Alexander Tekle <alexandertekle@google.com>
Reviewed-by: Marc Treib <treib@chromium.org>
Commit-Queue: Vipul Koul <koulvipul@google.com>
Cr-Commit-Position: refs/heads/main@{#1189297}