MaybeCommitResponse() can commit while State::kDataTransferStarted
Before this CL, once MaybeCommitResponse() is called, it sets
State::kResponseCommitted, and any states other than kResponseReceived
are guarded. This means the core logic of MaybeCommitResponse() is
executed only once.

However, when ServiceWorkerAutoPreload is enabled, this behavior is
problematic because AutoPreload may update the commit responsibility
from the fetch handler to the network request when fallback, and this
may happen while teeing the response data, which involves multiple
MaybeCommitResponse() calls. Actually this is very rare case but this is
the root cause of the test flakiness[1].

This CL fixes this issue by introducing the new state called
kDataTransferStarted. This is the State between kResponseReceived and
kResponseCommitted.

[1] https://ci.chromium.org/ui/p/chromium/builders/try/android-nougat-x86-rel/370942/overview

Change-Id: If58022934b764b00a97336226de7147a46c1d77c
Bug: 1472634
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4812929
Commit-Queue: Shunya Shishido <sisidovski@chromium.org>
Reviewed-by: Yoshisato Yanagisawa <yyanagisawa@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1189361}