Workaround: reset self update delay iff the registration is available.
In crbug.com/1475465, we observe the crash due to lack of the live
registration.  We believe that when there is a ServiceWorkerVersion,
it should always have an associated live registration.  However, the
linked crbug looks not and caused the crash due to null pointer
assignment.

We intend to duct-taping the code to mitigate the crash.  At the same
time, we wonder if it may bring a security issue.  Our conclusion is no.
Considering what this code do, it reset the update timer so that the
update of the ServiceWorker do not hit the timeout.  If the reset
happens where it should not be, update would never finish.  That is what
a CL that added the code originally wanted to do (See:
https://crrev.com/c/900763).  This CL makes the reset happen iff the live
registration is available, and it reduces the risk to keep the update
running.

Bug: 1475465
Change-Id: I04e3be21a3ed3fcdbab33c6e60fe44ca5191673c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4820729
Commit-Queue: Yoshisato Yanagisawa <yyanagisawa@chromium.org>
Reviewed-by: Minoru Chikamune <chikamune@chromium.org>
Reviewed-by: Shunya Shishido <sisidovski@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1189371}