[LanguagePacksHandwriting] Add IsHandwritingDlc
This is intended to be used to filter a list of DLCs that a user has
installed to only the relevant handwriting recognition ones.

This requires making `GetAllLanguagePackDlcIds()` in
language_pack_manager.cc public. Add documentation to the aforementioned
method explaining that it is internal only and will likely be removed in
the future.

Bug: b:294302472
Change-Id: I12606036bac02fda61069715fcd73a054649d18f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4776563
Reviewed-by: Claudio M <claudiomagni@chromium.org>
Commit-Queue: Michael Cui <mlcui@google.com>
Cr-Commit-Position: refs/heads/main@{#1189376}