[LanguagePacks] Add ComputeStringsDiff
This is intended for computing the DLCs to uninstall and install to get
to a desired state of DLCs. In combination with the functions
implemented in `handwriting.h`, this should allow us to trigger
installations and uninstallations automatically as detailed in the
attached bug.

Bug: b:294302472
Change-Id: I6e10e0c930f6b8517e622102416f586b925cb7e1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4780210
Commit-Queue: Michael Cui <mlcui@google.com>
Reviewed-by: Claudio M <claudiomagni@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1189386}