Add time parameter for AutofillPopupController::AcceptSuggestion.
This CL adds a parameter to AutofillPopupController::AcceptSuggestion
that describes the time at which the user event that caused the
decision to accept the Autofill suggestion was triggered.

As of now, most callers populate the field with base::TimeTicks::Now(),
but in principle this allows to pass the actual time when the event was
triggered by the user (as opposed to when it got processed by Chrome).
In cases of high latency between event triggering and event processing,
this becomes relevant.

Bug: 1475902, 1453815
Change-Id: Ie24198f53ae5d8df0b5124c24b70b2446a2883ad
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4811898
Reviewed-by: Bruno Braga <brunobraga@google.com>
Commit-Queue: Jan Keitel <jkeitel@google.com>
Reviewed-by: Vasilii Sukhanov <vasilii@chromium.org>
Reviewed-by: Christoph Schwering <schwering@google.com>
Cr-Commit-Position: refs/heads/main@{#1189392}