[ChromeOS] Add TrustedVault reauth dialog
Before: when user clicks on error notification, reauth page is opened
in a new browser tab. This is problematic in Ash, since after Lacros
launch there will be no browser (or Ash browser will be opened - but
this would be extremely confusing and should never happen).

After: the page is opened in WebUI dialog. This CL only makes this
change for Ash, but similar behavioral change will be made in Lacros
to keep the consistent.

The change is guarded by a dedicated feature flag and also by LacrosOnly
mode.

Bug: 1434656
Change-Id: Ic745617e23a2c09de123de684a54289e43e7998b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4790365
Reviewed-by: Mikel Astiz <mastiz@chromium.org>
Reviewed-by: Scott Violet <sky@chromium.org>
Commit-Queue: Maksim Moskvitin <mmoskvitin@google.com>
Reviewed-by: John Lee <johntlee@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1189393}