Remove StrictModeContext from the input Hot path on Android.
StrictModeContext was added due to developers using WebView encountering
the violation of a NonSDK usage. However we have noticed as this rolled
out that suppressing the error at run time results in a 1ms increase of
all input event handling at the median, and 2.1ms at the 95th
percentile.

This is too expensive to accept for input events which can come once
every 4ms on 120Hz devices. Thus reverse the suppression.

This means that on Android P to Android T (API level 28 to 33) this
strict mode violation will be reported, and down stream apps should add
it to their ignore list. In Android U this function was added to the
public API and thus the violation no longer occurs.

Bug: 1454299, b/280373224
Change-Id: I8b931af013a9a628488bfae156f076f38c59bdee
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4807915
Commit-Queue: Stephen Nusko <nuskos@chromium.org>
Reviewed-by: Omar Elmekkawy <mekk@chromium.org>
Reviewed-by: Avi Drissman <avi@chromium.org>
Auto-Submit: Stephen Nusko <nuskos@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1189400}