[iOS] Disable UNO flag for AdvancedSettingsSigninTestCase.
The AdvancedSettingsSigninCoordinator only creates the
ManageSyncSettingsCoordinator in the
SyncSettingsAccountState::kAdvancedInitialSyncSetup state.

All of its EG tests are written for the view in this state, which does
not exist in the UNO sign-in flows.
This CL disables the UNO flag kReplaceSyncPromosWithSignInPromos for
these EG tests.

Bug: 1462156
Change-Id: Ie330001510d6812feacae369a5f08e1162e54154
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4817049
Reviewed-by: Marc Treib <treib@chromium.org>
Reviewed-by: Jérôme Lebel <jlebel@chromium.org>
Commit-Queue: Jood Hajeer <jood@google.com>
Cr-Commit-Position: refs/heads/main@{#1189402}