Lacros proxy tracker: Send extension metadata on extension loaded
The Lacros proxy tracker sends updates to Ash with metadata describing
the extension which is controlling the proxy pref. Ash uses this
metadata to populate the network settings UI.

There are a few cases where the Lacros proxy tracker receives proxy pref
updates but the extension which is controlling the proxy is not yet
loaded:
- A new Chrome session is started, where the extension has set the proxy
pref in a previous session;
- Ash to Lacros profile migration;
- Disabling and then enabling an extension which is controlling the
proxy.

This happens because the ExtensionPrefs KeyedService is parsing the extension-set pref values from the user pref kExtensions and setting it into the extensions pref store before the extension are loaded. The user pref kExtensions is stored in the Preferences file and read from disk when the PrefService is created.

For the last two cases, Ash never receives the extension metadata.

This CL adds an observer for RegistryEvents in the Lacros proxy tracker
to retry sending the extension metadata when the extension is ready.

Bug: b/293845404, b/290184836
Change-Id: I2c221590da83ea9ef2cabcf38c4aadbb84e532b1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4748739
Reviewed-by: Yuta Hijikata <ythjkt@chromium.org>
Reviewed-by: Hidehiko Abe <hidehiko@chromium.org>
Commit-Queue: Andreea Costinas <acostinas@google.com>
Cr-Commit-Position: refs/heads/main@{#1189454}